requirements.txt
urllib3~=1.26.7
opensearch-py~=1.1.0
requests~=2.26.0
kubernetes~=18.20.0
retry~=0.9.2
waiting~=1.4.1
assisted-service-client~=2.1.0
sentry-sdk~=1.5
python-json-logger~=2.0
dataclasses~=0.8
wheel
smart-open[s3]~=5.2.1
boto3~=1.22.6
dpath~=2.0.6
kafka-python~=2.0.2
